# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Company - Service",
        "description": "Service documentation",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/institutions": {
            "post": {
                "tags": [
                    "Instituições"
                ],
                "summary": "Consulta instituições financeiras disponíveis",
                "description": "Retorna a lista de instituições financeiras disponíveis para um CPF específico.",
                "operationId": "29302464d6f2b1d5c22e0b1bbe8571ac",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cpf"
                                ],
                                "properties": {
                                    "cpf": {
                                        "type": "string",
                                        "example": "12345678901"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lista de instituições",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "instituicoes": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nome": {
                                                        "type": "string",
                                                        "example": "Banco XYZ"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "CPF inválido"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        },
        "/api/credit-offer": {
            "post": {
                "tags": [
                    "Ofertas de Crédito"
                ],
                "summary": "Buscar ofertas de crédito por CPF",
                "description": "Retorna as ofertas de crédito disponíveis para um CPF específico.",
                "operationId": "a023a195970f604bf19c88a099f3d521",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cpf"
                                ],
                                "properties": {
                                    "cpf": {
                                        "type": "string",
                                        "example": "12345678901"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lista de ofertas de crédito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "instituicoes": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nome": {
                                                        "type": "string",
                                                        "example": "Banco XYZ"
                                                    },
                                                    "modalidades": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "cod": {
                                                                    "type": "string",
                                                                    "example": "001"
                                                                },
                                                                "nome": {
                                                                    "type": "string",
                                                                    "example": "Crédito Pessoal"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "CPF inválido"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        },
        "/api/credit-simulation": {
            "post": {
                "tags": [
                    "Simulação de Crédito"
                ],
                "summary": "Simula uma oferta de crédito",
                "description": "Simula uma oferta de crédito para um CPF em uma instituição e modalidade específica.",
                "operationId": "e847cef53ecaa0a1ae47cfb4c2c0a611",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cpf",
                                    "institution_id",
                                    "codModalidade"
                                ],
                                "properties": {
                                    "cpf": {
                                        "type": "string",
                                        "example": "12345678901"
                                    },
                                    "institution_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "codModalidade": {
                                        "type": "string",
                                        "example": "001"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resultado da simulação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "valorMin": {
                                            "type": "number",
                                            "example": 5000
                                        },
                                        "valorMax": {
                                            "type": "number",
                                            "example": 8000
                                        },
                                        "jurosMes": {
                                            "type": "number",
                                            "example": 0.0495
                                        },
                                        "QntParcelaMin": {
                                            "type": "integer",
                                            "example": 12
                                        },
                                        "QntParcelaMax": {
                                            "type": "integer",
                                            "example": 48
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nenhuma simulação encontrada"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Instituições",
            "description": "Instituições"
        },
        {
            "name": "Ofertas de Crédito",
            "description": "Ofertas de Crédito"
        },
        {
            "name": "Simulação de Crédito",
            "description": "Simulação de Crédito"
        }
    ]
}